Finish bulk usage matching in csv, tsv or json. Move name interpretation into a reusable NameInterpreter class for matching services. Use IssueContainer instead of VerbatimRecord where possible.

+++ webservice/src/main/java/life/catalogue/WsServerConfig.java
@@ -117,7 +117,7 @@ public class WsServerConfig extends Configuration implements CorsBundleConfigura
    * The name parser timeout in milliseconds
    */
   @Min(100)
-  public int parserTimeout = 5000;
+  public long parserTimeout = 5000;
 
   /**
    * The maximum allowed time in seconds for a unix diff to take before throwing a time out.
